revert: move back InputLabel
